drivers: mmc: add functions to enable/disable aggressive clock gating
There are few places where we want to make sure that no clock gating takes
place. For example when we are updating several related fields in ios
structure and we don't want to accidentally pass the partially filled ios
to the host driver.

To solve this we add two functions to enable/disable the aggressive clock
gating where this is needed.

Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>
Cc: Linus Walleij <linus.walleij@linaro.org>